Poprawa ułożenia query do common
