+++ src/Controller/AudiobookController.php
@@ -11,8 +11,8 @@ use App\Model\Error\NotAuthorizeModel;
 use App\Model\Error\PermissionNotGrantedModel;
 use App\Model\User\AudiobookCoverModel;
 use App\Model\User\AudiobookCoversSuccessModel;
-use App\Query\User\AudiobookCoversQuery;
-use App\Query\User\AudiobookPartQuery;
+use App\Query\Common\AudiobookCoversQuery;
+use App\Query\Common\AudiobookPartQuery;
 use App\Repository\AudiobookRepository;
 use App\Service\AuthorizedUserServiceInterface;
 use App\Service\RequestServiceInterface;
