+++ src/Controller/NotificationController.php
@@ -12,8 +12,8 @@ use App\Model\Error\JsonDataInvalidModel;
 use App\Model\Error\NotAuthorizeModel;
 use App\Model\Error\PermissionNotGrantedModel;
 use App\Model\User\NotificationsSuccessModel;
-use App\Query\User\SystemNotificationActivateQuery;
-use App\Query\User\SystemNotificationQuery;
+use App\Query\Common\SystemNotificationActivateQuery;
+use App\Query\Common\SystemNotificationQuery;
 use App\Repository\NotificationCheckRepository;
 use App\Repository\NotificationRepository;
 use App\Service\AuthorizedUserServiceInterface;
