+++ src/Controller/RegisterController.php
@@ -12,8 +12,8 @@ use App\Exception\DataNotFoundException;
 use App\Exception\InvalidJsonDataException;
 use App\Model\Error\DataNotFoundModel;
 use App\Model\Error\JsonDataInvalidModel;
-use App\Query\User\RegisterConfirmSendQuery;
-use App\Query\User\RegisterQuery;
+use App\Query\Common\RegisterConfirmSendQuery;
+use App\Query\Common\RegisterQuery;
 use App\Repository\InstitutionRepository;
 use App\Repository\MyListRepository;
 use App\Repository\ProposedAudiobooksRepository;
