+++ src/Query/Common/AudiobookPartQuery.php
@@ -1,17 +1,22 @@
 <?php
 
-namespace App\Query\Admin;
+namespace App\Query\Common;
 
 use OpenApi\Attributes as OA;
 use Symfony\Component\Uid\Uuid;
 use Symfony\Component\Validator\Constraints as Assert;
 
-class AudiobookCommentGetQuery
+class AudiobookPartQuery
 {
     #[Assert\NotNull(message: "AudiobookId is null")]
     #[Assert\NotBlank(message: "AudiobookId is blank")]
     #[Assert\Uuid]
     private Uuid $audiobookId;
+
+    #[Assert\NotNull(message: "Part is null")]
+    #[Assert\NotBlank(message: "Part is empty")]
+    #[Assert\Type(type: "integer")]
+    private int $part;
 
     /**
      * @return Uuid
@@ -29,4 +34,21 @@ class AudiobookCommentGetQuery
     {
         $this->audiobookId = Uuid::fromString($audiobookId);;
     }
+
+    /**
+     * @return int
+     */
+    public function getPart(): int
+    {
+        return $this->part;
+    }
+
+    /**
+     * @param int $part
+     */
+    public function setPart(int $part): void
+    {
+        $this->part = $part;
+    }
+
 }
